refactor: change name of directories and files
